+++ InscryptionAPI/Card/CardModificationInfoManager.cs
@@ -1,17 +1,7 @@
-using APIPlugin;
 using DiskCardGame;
 using HarmonyLib;
-using InscryptionAPI.Guid;
-using InscryptionAPI.Helpers;
-using InscryptionAPI.Saves;
-using MonoMod.Cil;
 using Sirenix.Serialization.Utilities;
-using System.Collections;
-using System.Collections.ObjectModel;
-using System.Reflection;
-using System.Reflection.Emit;
 using System.Runtime.CompilerServices;
-using UnityEngine;
 
 namespace InscryptionAPI.Card;